Centralize production SQL for reuse by Hubble smoke tests
